psplash: delete conflicting psplash_git.bb
The angstrom layer and core layer both have a psplash_git.bb recipe.
The core layer recipe is preferable to use (as it can automatically
generate image data) and we have a bbappend which also patches the core
version.
Because the angstrom layer is listed in first in bblayers, it's recipe
takes precedence. Deleting the recipe from the angstrom layer a way around
this, as we otherwise want the angstrom layer to have higher importance.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>